Addressing most of the comments on PR.
The only one that left is adding more tests.
